+++ services/quiz/src/mock/db.ts
@@ -108,7 +108,7 @@ export const absentee = [
   ];
 
 export const studyInfo = {
-    studyInfo: 12,
+    studyId: 12,
     title: "아키텍쳐 크리너스너스너스너스너스",
     description: "동해 물과 백두산이 마르고 닳도록 하느님이 보우하사 우리나라 만세 무궁화 삼천리 화려강산 대한 사람 대한으로 길이 보전하세 동해 물과 백두산이 마르고 닳도록 하느님이 보우하사 우리나라 ",
     studyTags: ['북 스터디', '기술 스터디', 'JavaScript'],
